It took 3/4 of forever, but it finally builds/runs
Had to implement my own compiler for Tmod files...at least, it seemed
like I did, the builtin one kept crashing.  Will investigate that more
later.  Even with the custom compiler, I couldn't build from source;
the only way to get it to work at all was to precompile using xbuild
and have the tmod compiler target the Windows.dll and Mono.dll.
Also had to make sure to target framework v4.0, or it would
crash during load.  Learned how to use a single csproj file for
multiple build configurations, so that's something.

I'm honestly not sure if it will work on Windows, but I feel like it
should. It works on Linux, miraculously! The hotkeys seem to work like
they should (though the keybinds are messed up) and the proper
chest-to-player or vice-versa action is taken. The item sorting may not
be working, but I wasn't even expecting this much to work on a first
run, so I'm pretty pleased. I also couldn't figure out how to activate
chat to test the commands...but surely I'm just missing something.
Progress looks hopeful.