CSON experiment
Testing out storing Trait and Category definitions in .cson format (for
ease of writing and clarity of reading). These can be converted to json
later, or I could write a custom parser, or maybe they won't be used at
all, but will likely still be helpful for planning everything out.