Create ItemCategory class; impl merging
Also, moved Parser.cs to the Definitions directory with all of the
hjson files; this makes independent compilation and testing easier
since none of the things in that diretory rely on Terraria or
tModLoader or XNA or anything. The ItemCategory class was created
due to a need to store more information about a category than just
its list of requirements (for instance, whether it is a member of
a merged category).

Now, all the traits and category Definitions are loaded dynamically;
testing seems to indicate that all of that machinery is working
properly! It's time to finally start (re)implementing the sorting
algorithm (should be able to load dynamic sorting rules in much the
same manner, using the RuleBuilder) and then actually compile and test
it in game.