ItemInfo now includes method to retrieve Category
Using the Category tree to save a lot of unnecessary checks. The
determined category is cached on the ItemFlagInfo instance (I...think
that's persistent?), so the category lookup should only be done once
per created-item (and only when requested, which is during the sort/
stack operations). I'm not sure if it's going to problematic to
retrieve the item info for any-given item from a non-GlobalItem class;
if so, it may just be easier to ditch the ItemInfo stuff altogether and
maintain by own cache of item-type->Category mappings (since the category is
only based on the generic traits of an item and not anything unique to
an instance such as prefix). Once again, I guess we'll see when we
finally get to testing in-game!