Have to embed the hjson files, load w/ reflection
Took forever for me to figure out why it couldn't find the definition
files; apparently "<Content Include=" doesn't actually do anything with
the files...it just suggests to the program that those files are
available for use. So unless I wanted to distribute them alongside the
mod as a separate folder (which I might do, someday), they had to be
listed in the csproj as "<Embedded Include=" and loaded via reflection
rather than Directory access.

.Net has too many rules...and everyone relies on Visual Studio to
figure it all out and take care of it for them...makes it very
difficult for those of us outside of MSoft's greasy clutches to learn
anything about how to work with the framework...