Delay several actions until everything is properly loaded
It seems that the attempt to analyze loaded recipes was happening
before tModLoader had actually initialized any recipes; thus the
'material' and 'alchemy' categories were always empty. To avoid this,
recipe-analysis is now delayed until World-load.

Also, I entirely
disposed of ClassifiedItem, since, similarly, SetDefaults was being
called too early and not every category was properly set up yet. Now
each item is classified on the first attempt to access its category,
and the results of the classification and category-check are cached
for the item type. I've not noticed any performance issues with this so
far; in fact, even when sorting for the first time, there doesn't seem
to be even a single hiccup or skipped frame. So far so good.