Add Tool subcats; start adding sorting rules
The rules aren't parsed yet. Additionally, testing indicates that there
likely needs to be a separation between "order in which categories are
checked for a match" and "order in which categories are sorted", like
there was in the old TIH. Otherwise, items will probably keep landing
in the wrong categories and won't sort correctly.