Generics don't like me; Category is Item-specific
that's what I get for trying to be idealistic and build a
'universal' solution. But, now ItemCategory can be used as an IComparer
for sorting items matching that category.

So, after plugging in all the machinery to make that happen--sorting
still works, much like it used to. The ordering looks to be a bit
different, but since so many items are still not quite falling in the
correct categories, it's hard to tell how well its working. But not
being totally broken is a relief, so next step is tightening up the
categories.