Replace shorts with ints
Since we're no longer doing any fancy bitshifting to get the ordinals
or priorities or anything, using shorts is no longer necessary. It
would save a bit of memory (barely), but as the language and the CPU is
optimized for 32-bit integer calculations, using ints should give a bit
(ha!) of a performance boost.

On a related note, due to the new priority-mechanics, I removed the
priority value from most of the category definitions; their order of
appearance in the definition files now defines their sorting.