Try another attempt at making categories generic
I feel like this is important...for some reason. For the future, maybe.
For posterity. For the world we leave behind.

Or maybe I'm just being overly-concerned about the 'right' (fancy?) way
to do things. I do feel that if I'm going to improve my skills, I
should be able to right reusable, generic code like this. And it seems
that this attempt works, so maybe I am improving.

Well, by "works", I mean "totally breaks sorting," but I think that the
issue behind that is elsewhere. Hopefully.