Sets now check proper Item field automagically
This was actually in the last commit, though it wasn't supposed to
be...whoops. Anyway, The "InvisibleHand.Items.Set" class was made
generic, and I implemented two special subclasses ("TileIDSet" and
"ItemIDSet") that can be called with "set.Contains(myItem)" and the set
will check that it contains either myItem.createTile or myItem.type as
appropriate given its subclass type. Saves some trouble figuring out
which field to use when writing rules and trait-checks.