Allow all categories to be disabled
This requires re-parenting a category if its parent has been disabled.

There's now an ActiveCategories set (of category Ids) for tracking which
categories are currently enabled. Additionally, Union Categories are
never added to this set, which means they are no longer present in the
category tree. Instead, the tree will be checked as if the members of a
union were in their original locations; if a member reports a match, it
will report its Union as the matched category, and the union will be
used to determine where to sort the matched item.