Improve Union category handling.
Enforce a single union per category when attempting to add a member to
a union, unless a value of 'true' is passed for the force_remove
parameter. Also, fixed the infinite recursion when comparing items in a
merged category, and without having to perform an entire Match check on
both items.

On a different note, spears are still not showing as spears...just as
weapon.melee. Don't know what's up with that.