Actually fix spears (and boomerangs, and more)
So, it turns out that Projectiles don't exist until they, well--exist.
As in, none of the properties of a projectile object are initialized
until the projectile is about to be used and SetDefaults() is called.
Thus, none of my checks that rely on checking properties of the item's
projectile (such as aiStyle) were returning true because no aiStyle had
been set. To get around this, I reference the projectile object and
call SetDefaults on it before checking the AI. It seems that this
doesn't actually make a projectile appear or anything (it just sets
values), so hopefully this won't mess anything up or cause a memory
leak, etc. Basic testing seemed to indicate all is well, but I'll keep
a look out.