Add type code for sub-category definition
Rather than relying on the name starting with '@'. Also, category
names/labels must now be surrounded by double quotes; this makes it
easier to allow almost any arbitrary character as part of a category
name. Finally, the separator between each section of a category
definition can now be either a colon or semicolon; it doesn't matter
which, nor is consistency required.